Refactor agent.py to separate files
Create common.py, nshot.py, cot.py, react.py in src/phantom_eval/agents/
from parts of old src/phantom_eval/agent.py
