+++ src/phantom_eval/agents/common.py
@@ -1,0 +1,184 @@
+import abc
+import subprocess
+from collections import Counter
+
+import openai
+import pandas as pd
+from langchain_community.vectorstores import FAISS
+from langchain_core.embeddings import Embeddings
+
+from phantom_eval._types import Conversation, LLMChatResponse
+from phantom_eval.gpu_utils import get_gpu_count
+from phantom_eval.llm.common import InferenceGenerationConfig, LLMChat, aggregate_usage
+from phantom_eval.prompts import LLMPrompt
+from phantom_eval.score import normalize_pred
+
+
+class Agent(abc.ABC):
+    """
+    Abstract class for an agent that implements an evaluation method by prompting an LLM.
+    """
+
+    def __init__(self, text_corpus: pd.DataFrame, llm_prompt: LLMPrompt):
+        """
+        Args:
+            text_corpus (pd.DataFrame): The text corpus to search for answers.
+                Must contain two columns: 'title' and 'article'.
+            llm_prompt (LLMPrompt): The prompt to be used by the agent.
+        """
+        self.text_corpus = text_corpus
+        self.llm_prompt = llm_prompt
+        self.agent_interactions: Conversation | list[Conversation] = None
+
+    @abc.abstractmethod
+    def run(
+        self, llm_chat: LLMChat, question: str, inf_gen_config: InferenceGenerationConfig
+    ) -> LLMChatResponse:
+        """
+        Run the agent with an LLM on a given question.
+        """
+
+    @abc.abstractmethod
+    async def batch_run(
+        self, llm_chat: LLMChat, questions: list[str], inf_gen_config: InferenceGenerationConfig
+    ) -> list[LLMChatResponse]:
+        """
+        Asynchronously run the agent with an LLM on a list of questions.
+        """
+
+    @abc.abstractmethod
+    def _build_agent_prompt(self, question: str) -> str:
+        """
+        Returns the agent prompt with the given question.
+        The prompt may depend on the agent's internal state.
+        """
+
+    def reset(self) -> None:
+        """
+        Reset the agent to its initial state.
+        """
+
+
+class SCMixin:
+    def __init__(self, num_votes: int, sep: str):
+        """
+        Args:
+            num_votes (int): The number of votes to take for the majority vote.
+            sep (str): The separator used to split the prediction.
+        """
+        self.num_votes = num_votes
+        self.sep = sep
+
+    def take_majority_vote(self, responses: list[LLMChatResponse], sep: str) -> LLMChatResponse:
+        """
+        Take the majority vote over all answers from the response predictions.
+
+        Args:
+            responses (list[LLMChatResponse]): List of response predictions.
+                Each response pred may contain multiple answers e.g. A, B, C.
+                So response preds can be like [[A<sep>B], [A<sep>B<sep>C]] where [A<sep>B] is the first
+                response pred and [A<sep>B<sep>C] is the second response pred.
+            sep (str): The separator used to split the prediction.
+
+        Returns:
+            LLMChatResponse: the majority vote as a single string of answers separated by <sep>
+                (the output string is in LLMChatResponse.pred).
+        """
+        n_preds = len(responses)
+        preds: list[set[str]] = [normalize_pred(response.pred, sep) for response in responses]
+        total_usage: dict = aggregate_usage([response.usage for response in responses])
+
+        # Flatten the list of sets to a single list, e.g. becomes [A, B, A, B, C]
+        all_answers: list[str] = [answer for pred in preds for answer in pred]
+        vote_counts = Counter(all_answers)
+
+        # Select all answers that have more than n_preds / 2 counts
+        majority_responses = [answer for answer, count in vote_counts.items() if count > n_preds / 2]
+        error = (
+            None
+            if len(majority_responses) > 0
+            else f"<agent_error>No majority vote found in {vote_counts=}.</agent_error>"
+        )
+
+        majority_responses_str = sep.join(majority_responses)
+        return LLMChatResponse(pred=majority_responses_str, usage=total_usage, error=error)
+
+
+class CustomEmbeddings(Embeddings):
+    def __init__(self, client):
+        self.client = client
+        self.model = self.client.models.list().data[0].id
+
+    def embed_documents(self, texts: list[str]) -> list[list[float]]:
+        """Embed search docs."""
+        return [obj.embedding for obj in self.client.embeddings.create(input=texts, model=self.model).data]
+
+    def embed_query(self, text: str) -> list[float]:
+        """Embed query text."""
+        return self.embed_documents([text])[0]
+
+
+class RAGMixin:
+    def __init__(
+        self,
+        text_corpus: pd.DataFrame,
+        embedding_model_name: str = "whereisai/uae-large-v1",
+        retriever_num_documents: int = 4,
+        tensor_parallel_size: int | None = 1,
+        port: int = 8001,
+    ):
+        """
+        Args:
+            embedding_model_name (str): The embedding method for the vector database.
+                Defaults to WhereIsAI/UAE-Code-Large-V. All VLLM models are supported.
+            retriever_num_documents (int): Number of documents retrieved.
+                Defaults to 4. See
+                https://api.python.langchain.com/en/latest/vectorstores/langchain_community.vectorstores.
+                faiss.FAISS.html#langchain_community.vectorstores.faiss.FAISS.as_retriever for other options.
+        """
+
+        self.embedding_model_name = embedding_model_name
+        texts = self.text_corpus["article"].tolist()
+
+        # launch server
+        subprocess.call(
+            [
+                "./src/phantom_eval/launch_embedding_server.sh",
+                embedding_model_name,
+                str(port),
+                str(get_gpu_count() - 1),
+            ]
+        )
+
+        # build retriever
+        BASE_URL = f"http://0.0.0.0:{port}/v1"
+        API_KEY = "token-abc123"
+        client = openai.OpenAI(
+            base_url=BASE_URL,
+            api_key=API_KEY,
+        )
+        embeddings = CustomEmbeddings(client)
+        vectorstore = FAISS.from_texts(texts, embeddings)
+        self.retriever = vectorstore.as_retriever(search_kwargs={"k": retriever_num_documents})
+
+    def get_RAG_evidence(self, question: str) -> str:
+        """
+        Returns retrieved articles in the text corpus as evidence.
+        """
+        self.format_RAG_docs = lambda docs: "\n================\n\n".join(doc.page_content for doc in docs)
+        evidence = self.format_RAG_docs(self.retriever.invoke(question))
+        return evidence
+
+
+REASONING_LLM_NAMES: list[str] = [
+    "deepseek-ai/deepseek-r1-distill-qwen-32b",
+    "deepseek-ai/deepseek-r1-distill-qwen-7b",
+    "deepseek-ai/deepseek-r1-distill-qwen-1.5b",
+]
+
+
+def _get_evidence(text_corpus: pd.DataFrame) -> str:
+    """Utility for constructing evidence
+    Returns all articles (concatenated as a string) in the text corpus as evidence.
+    """
+    return "\n================\n\n".join(text_corpus["article"])
